Pull just open merges from gitlab instead of first page of all of them
getMergeRequests returns only the first page (20) of the merge requests,
which means most of the time the one related to the push won't be found
getOpenMergeRequests return all open merge requests, which not only
results in smaller array, but also should guarantee a match (if there
should be one).

Unfortunately the java-git api does not provide a overload with just
projectId, though the extra instance() call.